Fix Unicode Error in hints section
This commit fixes the error that was brought to our attention by
our automated code diagnostics tool. The problem was successfully
replicated by using the unicode character: 'μ' in the hints section of
the xblock and attempting to save changes as an instructor. The problem was
that we were previously calling python's 'format' method on a
non-unicode string, but passing in one or more unicode arguments to the
function. In order to fix the bug, I ensured that format was being called on
a unicode string. For more details on the nature of the issue see:
http://www.saltycrane.com/blog/2014/07/old-string-formatting-and-new-format-method-handle-unicode-differently/