Add tox dependency on opaque-keys to match platform
This commit adds a dependency to the tox setup on
the same version of opaque_keys used by
Stanford-Online's version of edx-platform. This
was added as a tox dependency instead of being
included in the `install_requires` array of the
xblock's setup.py because the version of opaque_keys
is not available on PyPi.